Forgot this little one.
Note: after a while, the K+up; K+left method is quite convenient, tell
me otherwise. Maybe we should store the previous commands for each
window into something, so that we don't rely on the detection of the
position/size by xdotool (that can fail if the window is non freely
resizeable and that is buggy when the borders are too thick.